Added permission system and switched to new infrastructure

+++ src/Spirebyte.Services.Issues.Application/Commands/Handlers/DeleteIssueHandler.cs
@@ -1,10 +1,12 @@
 ﻿using Convey.CQRS.Commands;
 using Microsoft.Extensions.Logging;
+using Spirebyte.Services.Issues.Application.Clients.Interfaces;
+using Spirebyte.Services.Issues.Application.Events;
 using Spirebyte.Services.Issues.Application.Exceptions;
+using Spirebyte.Services.Issues.Application.Services.Interfaces;
+using Spirebyte.Services.Issues.Core.Constants;
 using Spirebyte.Services.Issues.Core.Repositories;
 using System.Threading.Tasks;
-using Spirebyte.Services.Issues.Application.Events;
-using Spirebyte.Services.Issues.Application.Services.Interfaces;
 
 namespace Spirebyte.Services.Issues.Application.Commands.Handlers
 {
@@ -13,12 +15,16 @@ namespace Spirebyte.Services.Issues.Application.Commands.Handlers
         private readonly IIssueRepository _issueRepository;
         private readonly ILogger<DeleteIssueHandler> _logger;
         private readonly IMessageBroker _messageBroker;
+        private readonly IProjectsApiHttpClient _projectsApiHttpClient;
+        private readonly IAppContext _appContext;
 
-        public DeleteIssueHandler(IIssueRepository issueRepository, ILogger<DeleteIssueHandler> logger, IMessageBroker messageBroker)
+        public DeleteIssueHandler(IIssueRepository issueRepository, ILogger<DeleteIssueHandler> logger, IMessageBroker messageBroker, IProjectsApiHttpClient projectsApiHttpClient, IAppContext appContext)
         {
             _issueRepository = issueRepository;
             _logger = logger;
             _messageBroker = messageBroker;
+            _projectsApiHttpClient = projectsApiHttpClient;
+            _appContext = appContext;
         }
 
         public async Task HandleAsync(DeleteIssue command)
@@ -27,6 +33,11 @@ namespace Spirebyte.Services.Issues.Application.Commands.Handlers
             if (issue is null)
             {
                 throw new IssueNotFoundException(command.IssueId);
+            }
+
+            if (!await _projectsApiHttpClient.HasPermission(IssuePermissionKeys.DeleteIssues, _appContext.Identity.Id, issue.ProjectId))
+            {
+                throw new ActionNotAllowedException();
             }
 
             await _issueRepository.DeleteAsync(issue.Id);
